Replace AbstractGameObject with GameObject
This interface can be extended by Player, making it possible to
treat Player as a GameObject throughout the code.

< interface, naming
