+++ src/main/java/nl/tudelft/controller/Modifiable.java
@@ -1,6 +1,6 @@
 package nl.tudelft.controller;
 
-import nl.tudelft.model.AbstractGameObject;
+import nl.tudelft.model.GameObject;
 
 /**
  * The Modifiable interface should be implemented by the Level object, allowing other objects to
@@ -18,7 +18,7 @@ public interface Modifiable {
      *            GameObject - The object that the caller wants added.
      * @return boolean - true if the addition was successful, false if it was not.
      */
-    public void toAdd(AbstractGameObject obj);
+    public void toAdd(GameObject obj);
 
     /**
      * Removes an object from the class, provided that it is in there.
@@ -28,6 +28,6 @@ public interface Modifiable {
      * @return boolean - true if the object is no longer in the class, whether it was removed by
      *         this call or already absent.
      */
-    public void toRemove(AbstractGameObject obj);
+    public void toRemove(GameObject obj);
 
 }
