+++ src/main/java/nl/tudelft/controller/collision/CollisionInteractionMap.java
@@ -5,18 +5,24 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import nl.tudelft.model.AbstractGameObject;
 import nl.tudelft.model.Game;
+import nl.tudelft.model.GameObject;
 
 /**
  * This class was inspired by CollisionInteractionMap of jpacman.
  *
- * <p>The different lies in the way handlers are matched to two classes.
- * Our version doesn't stop at the first match, it recursively walks through the
- * hierarchy and calls every handler that is matched.</p>
+ * <p>
+ * The different lies in the way handlers are matched to two classes. Our version doesn't stop at
+ * the first match, it recursively walks through the hierarchy and calls every handler that is
+ * matched.
+ * </p>
  *
- * <p>original comment:</p>
- * <p>A map of possible collisions and their handlers.</p>
+ * <p>
+ * original comment:
+ * </p>
+ * <p>
+ * A map of possible collisions and their handlers.
+ * </p>
  *
  * @author Michael de Jong
  * @author Jeroen Roosen
@@ -28,9 +34,8 @@ public class CollisionInteractionMap implements CollisionMap {
     /**
      * The collection of collision handlers.
      */
-    private final Map<
-            Class<? extends AbstractGameObject>,
-            Map<Class<? extends AbstractGameObject>, CollisionHandler<?, ?>>> handlers;
+    private final Map<Class<? extends GameObject>,
+        Map<Class<? extends GameObject>, CollisionHandler<?, ?>>> handlers;
 
     /**
      * Creates a new, empty collision map.
@@ -40,8 +45,8 @@ public class CollisionInteractionMap implements CollisionMap {
     }
 
     /**
-     * Adds a two-way collision interaction to this collection, i.e. the
-     * collision handler will be used for both C1 versus C2 and C2 versus C1.
+     * Adds a two-way collision interaction to this collection, i.e. the collision handler will be
+     * used for both C1 versus C2 and C2 versus C1.
      *
      * @param <C1>
      *            The collider type.
@@ -54,9 +59,8 @@ public class CollisionInteractionMap implements CollisionMap {
      * @param handler
      *            The handler that handles the collision.
      */
-    public <C1 extends AbstractGameObject, C2 extends AbstractGameObject> void onCollision(
-            Class<C1> collider, Class<C2> collidee,
-            CollisionHandler<C1, C2> handler) {
+    public <C1 extends GameObject, C2 extends GameObject> void onCollision(Class<C1> collider,
+            Class<C2> collidee, CollisionHandler<C1, C2> handler) {
         onCollision(collider, collidee, true, handler);
     }
 
@@ -72,15 +76,13 @@ public class CollisionInteractionMap implements CollisionMap {
      * @param collidee
      *            The collidee type.
      * @param symetric
-     *            <code>true</code> if this collision is used for both
-     *            C1 against C2 and vice versa;
-     *            <code>false</code> if only for C1 against C2.
+     *            <code>true</code> if this collision is used for both C1 against C2 and vice
+     *            versa; <code>false</code> if only for C1 against C2.
      * @param handler
      *            The handler that handles the collision.
      */
-    public <C1 extends AbstractGameObject, C2 extends AbstractGameObject> void onCollision(
-            Class<C1> collider, Class<C2> collidee, boolean symetric,
-            CollisionHandler<C1, C2> handler) {
+    public <C1 extends GameObject, C2 extends GameObject> void onCollision(Class<C1> collider,
+            Class<C2> collidee, boolean symetric, CollisionHandler<C1, C2> handler) {
         addHandler(collider, collidee, handler);
         if (symetric) {
             addHandler(collidee, collider, new InverseCollisionHandler<>(handler));
@@ -97,21 +99,19 @@ public class CollisionInteractionMap implements CollisionMap {
      * @param handler
      *            The handler that handles the collision.
      */
-    private void addHandler(Class<? extends AbstractGameObject> collider,
-                            Class<? extends AbstractGameObject> collidee,
-                            CollisionHandler<?, ?> handler) {
+    private void addHandler(Class<? extends GameObject> collider,
+            Class<? extends GameObject> collidee, CollisionHandler<?, ?> handler) {
         if (!handlers.containsKey(collider)) {
             handlers.put(collider, new HashMap<>());
         }
 
-        Map<Class<? extends AbstractGameObject>, CollisionHandler<?, ?>> map = handlers
-                .get(collider);
+        Map<Class<? extends GameObject>, CollisionHandler<?, ?>> map = handlers.get(collider);
         map.put(collidee, handler);
     }
 
     /**
-     * Handles the collision between two colliding parties, all suitable
-     * collision handlers are used.
+     * Handles the collision between two colliding parties, all suitable collision handlers are
+     * used.
      *
      * @param <C1>
      *            The collider type.
@@ -124,22 +124,20 @@ public class CollisionInteractionMap implements CollisionMap {
      */
     @SuppressWarnings("unchecked")
     @Override
-    public <C1 extends AbstractGameObject, C2 extends AbstractGameObject> void collide(
-            Game game,
-            C1 collider,
+    public <C1 extends GameObject, C2 extends GameObject> void collide(Game game, C1 collider,
             C2 collidee) {
-        List<Class<? extends AbstractGameObject>> colliderInheritance =
+        List<Class<? extends GameObject>> colliderInheritance =
                 getInheritance(collider.getClass());
-        List<Class<? extends AbstractGameObject>> collideeInheritance =
+        List<Class<? extends GameObject>> collideeInheritance =
                 getInheritance(collidee.getClass());
 
-        for (Class<? extends AbstractGameObject> colliderKey : colliderInheritance) {
+        for (Class<? extends GameObject> colliderKey : colliderInheritance) {
             if (handlers.get(colliderKey) != null) {
-                Map<Class<? extends AbstractGameObject>, CollisionHandler<?, ?>> colliderMap =
+                Map<Class<? extends GameObject>, CollisionHandler<?, ?>> colliderMap =
                         handlers.get(colliderKey);
-                for (Class<? extends AbstractGameObject> collideeKey : collideeInheritance) {
+                for (Class<? extends GameObject> collideeKey : collideeInheritance) {
                     if (colliderMap.get(collideeKey) != null) {
-                        ((CollisionHandler<C1, C2>)colliderMap.get(collideeKey))
+                        ((CollisionHandler<C1, C2>) colliderMap.get(collideeKey))
                                 .handleCollision(game, collider, collidee);
                     }
                 }
@@ -151,26 +149,25 @@ public class CollisionInteractionMap implements CollisionMap {
      * Returns a list of all classes and interfaces the class inherits.
      *
      * @param clazz
-     *            The class to create a list of super classes and interfaces
-     *            for.
+     *            The class to create a list of super classes and interfaces for.
      * @return A list of all classes and interfaces the class inherits.
      */
     @SuppressWarnings("unchecked")
-    private List<Class<? extends AbstractGameObject>> getInheritance(
-            Class<? extends AbstractGameObject> clazz) {
-        List<Class<? extends AbstractGameObject>> found = new ArrayList<>();
+    private List<Class<? extends GameObject>>
+            getInheritance(Class<? extends GameObject> clazz) {
+        List<Class<? extends GameObject>> found = new ArrayList<>();
         found.add(clazz);
 
         int index = 0;
         while (found.size() > index) {
             Class<?> current = found.get(index);
             Class<?> superClass = current.getSuperclass();
-            if (superClass != null && AbstractGameObject.class.isAssignableFrom(superClass)) {
-                found.add((Class<? extends AbstractGameObject>) superClass);
+            if (superClass != null && GameObject.class.isAssignableFrom(superClass)) {
+                found.add((Class<? extends GameObject>) superClass);
             }
             for (Class<?> classInterface : current.getInterfaces()) {
-                if (AbstractGameObject.class.isAssignableFrom(classInterface)) {
-                    found.add((Class<? extends AbstractGameObject>) classInterface);
+                if (GameObject.class.isAssignableFrom(classInterface)) {
+                    found.add((Class<? extends GameObject>) classInterface);
                 }
             }
             index++;
@@ -188,9 +185,7 @@ public class CollisionInteractionMap implements CollisionMap {
      * @param <C2>
      *            The collidee type.
      */
-    private static class InverseCollisionHandler<
-            C1 extends AbstractGameObject,
-            C2 extends AbstractGameObject>
+    private static class InverseCollisionHandler<C1 extends GameObject, C2 extends GameObject>
             implements CollisionHandler<C1, C2> {
 
         /**
@@ -209,8 +204,8 @@ public class CollisionInteractionMap implements CollisionMap {
         }
 
         /**
-         * Handles this collision by flipping the collider and collidee, making
-         * it compatible with the initial collision.
+         * Handles this collision by flipping the collider and collidee, making it compatible with
+         * the initial collision.
          */
         @Override
         public void handleCollision(Game game, C1 collider, C2 collidee) {
