+++ src/test/java/nl/tudelft/model/pickups/powerup/MoneyPowerupTest.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import nl.tudelft.controller.resources.ResourcesWrapper;
+import nl.tudelft.model.player.ConcretePlayer;
 import nl.tudelft.model.player.Player;
 
 import org.junit.Test;
@@ -31,7 +32,7 @@ public class MoneyPowerupTest {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         MoneyPowerup powerup = new MoneyPowerup(mockedResources, 0, 0);
 
-        Player player = new Player(mockedResources, 0, 0, true);
+        Player player = new ConcretePlayer(mockedResources, 0, 0, true);
 
         assertEquals(0, player.getMoney());
         assertFalse(powerup.isActive());
@@ -48,7 +49,7 @@ public class MoneyPowerupTest {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         MoneyPowerup powerup = new MoneyPowerup(mockedResources, 0, 0);
 
-        Player player = new Player(mockedResources, 0, 0, true);
+        Player player = new ConcretePlayer(mockedResources, 0, 0, true);
 
         assertEquals(0, player.getMoney());
         assertFalse(powerup.isActive());
