+++ src/test/java/nl/tudelft/model/pickups/powerup/ShieldPowerupTest.java
@@ -13,6 +13,7 @@ import static org.mockito.Mockito.when;
 import nl.tudelft.controller.Modifiable;
 import nl.tudelft.controller.resources.ResourcesWrapper;
 import nl.tudelft.controller.util.SemRectangle;
+import nl.tudelft.model.player.ConcretePlayer;
 import nl.tudelft.model.player.Player;
 
 import org.junit.Test;
@@ -22,79 +23,79 @@ import org.newdawn.slick.Image;
 import org.newdawn.slick.SlickException;
 
 public class ShieldPowerupTest {
-    
+
     @Test
     public void testConstructor() {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         Image mockedImg = mock(Image.class);
         when(mockedResources.getPickupPowerShield()).thenReturn(mockedImg);
         ShieldPowerup powerup = new ShieldPowerup(mockedResources, 0, 0);
-        
+
         assertEquals(mockedResources.getPickupPowerShield(), powerup.getImage());
         assertEquals(0.0f, powerup.getLocX(), 0.0f);
         assertEquals(0.0f, powerup.getLocY(), 0.0f);
     }
-    
+
     @Test
     public void testActivate1() {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         ShieldPowerup powerup = new ShieldPowerup(mockedResources, 0, 0);
-        
-        Player player = new Player(mockedResources, 0, 0, true);
-        
+
+        Player player = new ConcretePlayer(mockedResources, 0, 0, true);
+
         assertFalse(player.hasPowerup(Powerup.SHIELD));
         assertFalse(powerup.isActive());
         assertFalse(player.hasPowerup(Powerup.INVINCIBLE));
-        
+
         powerup.activate(player);
-        
+
         assertEquals(powerup, player.getPowerup(Powerup.SHIELD));
         assertTrue(powerup.isActive());
     }
-    
+
     @Test
     public void testActivate2() {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         ShieldPowerup powerup = new ShieldPowerup(mockedResources, 0, 0);
-        
-        Player player = new Player(mockedResources, 0, 0, true);
+
+        Player player = new ConcretePlayer(mockedResources, 0, 0, true);
         InvinciblePowerup invincibleMocked = mock(InvinciblePowerup.class);
-        
+
         assertFalse(player.hasPowerup(Powerup.SHIELD));
         assertFalse(powerup.isActive());
-        
+
         player.setPowerup(Powerup.INVINCIBLE, invincibleMocked);
         assertTrue(player.hasPowerup(Powerup.INVINCIBLE));
-        
+
         powerup.activate(player);
-        
+
         assertFalse(player.hasShield());
         assertTrue(powerup.isActive());
         assertTrue(powerup.willBeRemoved());
     }
-    
+
     @Test
     public void testActivate3() {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         ShieldPowerup powerup = new ShieldPowerup(mockedResources, 0, 0);
-        
-        Player player = new Player(mockedResources, 0, 0, true);
-        
+
+        Player player = new ConcretePlayer(mockedResources, 0, 0, true);
+
         assertFalse(player.hasPowerup(Powerup.SHIELD));
         assertFalse(powerup.isActive());
         assertFalse(player.hasPowerup(Powerup.INVINCIBLE));
-        
+
         powerup.activate(player);
-        
+
         assertEquals(powerup, player.getPowerup(Powerup.SHIELD));
         assertTrue(powerup.isActive());
-        
+
         ShieldPowerup powerup2 = new ShieldPowerup(mockedResources, 0, 0);
-        
+
         assertTrue(player.hasPowerup(Powerup.SHIELD));
-        
+
         powerup2.activate(player);
-        
+
         assertEquals(powerup2, player.getPowerup(Powerup.SHIELD));
         assertTrue(powerup2.isActive());
         assertTrue(powerup.willBeRemoved());
@@ -105,7 +106,7 @@ public class ShieldPowerupTest {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         Hit3ShieldPowerup powerup = new Hit3ShieldPowerup(mockedResources, 0, 0);
 
-        Player player = new Player(mockedResources, 0, 0, true);
+        Player player = new ConcretePlayer(mockedResources, 0, 0, true);
 
         assertFalse(player.hasPowerup(Powerup.SHIELD));
         assertFalse(powerup.isActive());
@@ -124,44 +125,44 @@ public class ShieldPowerupTest {
         assertEquals(powerup, player.getPowerup(Powerup.SHOPSHIELD));
         assertTrue(powerup2.willBeRemoved());
     }
-    
+
     @Test
     public void testUpdate1() throws SlickException {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         ShieldPowerup powerup = new ShieldPowerup(mockedResources, 0, 0);
-        
+
         assertFalse(powerup.isHit());
         assertEquals(0, powerup.getRemovalCounter());
-        
+
         Modifiable mockedContainer = mock(Modifiable.class);
-        
+
         powerup.update(mockedContainer, 0);
-        
+
         assertEquals(0, powerup.getRemovalCounter());
-        
+
         powerup.setHit(true);
         powerup.update(mockedContainer, 0);
         assertEquals(1, powerup.getRemovalCounter());
     }
-    
+
     @Test
     public void testUpdate2() throws SlickException {
         ResourcesWrapper mockedResources = mock(ResourcesWrapper.class);
         ShieldPowerup powerup = new ShieldPowerup(mockedResources, 0, 0);
-        Player player = new Player(mockedResources, 0, 0, true);
-        
+        Player player = new ConcretePlayer(mockedResources, 0, 0, true);
+
         powerup.activate(player);
-        
+
         assertFalse(powerup.isHit());
         assertEquals(0, powerup.getRemovalCounter());
-        
-        powerup.setRemovalCounter(120);        
+
+        powerup.setRemovalCounter(120);
         Modifiable mockedContainer = mock(Modifiable.class);
-        
+
         powerup.update(mockedContainer, 0);
-        
+
         assertEquals(120, powerup.getRemovalCounter());
-        
+
         assertTrue(powerup.willBeRemoved());
         assertFalse(player.hasShield());
         assertFalse(player.hasPowerup(Powerup.SHIELD));
